Fix clippy warnings

+++ tests/v0_9_0/data_model_map.rs
@@ -16,7 +16,7 @@ fn test() -> anyhow::Result<()> {
                     (
                         k,
                         Value {
-                            value_type: Some(ValueType::IntegerValue(i64::from(v))),
+                            value_type: Some(ValueType::IntegerValue(v)),
                         },
                     )
                 })
