+++ tests/v0_9_0/data_model_struct_variant.rs
@@ -27,7 +27,7 @@ fn test() -> anyhow::Result<()> {
                             (
                                 k,
                                 Value {
-                                    value_type: Some(ValueType::IntegerValue(i64::from(v))),
+                                    value_type: Some(ValueType::IntegerValue(v)),
                                 },
                             )
                         })
